Func implementation

+++ block.h
@@ -6,8 +6,8 @@
 #ifndef BLOCK
 #define BLOCK
 class Block{
-    int block_id;
-    int coin_number;
+    size_t block_id;
+    size_t coin_number;
     std::string sender, recipient;
     std::string block_hash;
     time_t create_time;
